test all metas for pair for file table
